Extended Probability Distribution Functions

The package provides a consistent, unified and extensible framework for parameter estimation of probability distributions; it extends parameter estimation procedures to allow for weighted samples; moreover, it extends the gallery of available distributions.

Details

Package:	ExtDist

Version:	0.6

License:	GPL (>= 2)

Imports:	numDeriv, optimx, VGAM, SuppDists, truncdist

Suggests:	knitr, ggplot2, xtable, PerformanceAnalytics

Roxygen:	list(wrap = FALSE)

VignetteBuilder:	knitr

Built:	R 3.1.1; ; 2015-03-31 03:52:27 UTC; unix

Index:

Beta                    The standard Beta Distribution.
Beta_ab                 The four-Parameter beta Distribution.
Burr                    The Burr's Distribution.
DistSelCriteriaValues   Distribution Selection Criteria Values.
ExtDist-package         Extended Probability Distribution Functions
Gamma                   The Gamma Distribution.
Gumbel                  The Gumbel Distribution.
JohnsonSB               The Johnson SB Distribution.
JohnsonSU               The Johnson SU Distribution.
Laplace                 The Laplace Distribution.
Logistic                The Logistic Distribution.
Normal                  The Normal Distribution.
Normal_sym_trunc_ab     The symmetric truncated normal distribution.
Normal_trunc_ab         The truncated normal distribution.
SRTB_ab                 The Symmetric-Reflected Truncated Beta (SRTB)
                        Distribution.
SSRTB                   The standard Symmetric-Reflected Truncated Beta
                        (SRTB) Distribution.
Triangular              The Triangular Distribution.
Uniform                 The Uniform Distribution.
Weibull                 The Weibull Distribution.
bestDist                Best distribution for (weighted) sample.
compareDist             Compare sample and fitted distributions
eDist                   S3 methods for manipulating eDist objects.
eval.estimation         Parameter Estimation Evaluation.
wmle                    Weighted Maximum Likelihood Estimation.

Further information is available in the following vignettes:

Distributions-Beta	Distributions-Beta (source)

Distributions-Normal	Distributions-Normal (source)

Distributions	Distributions-Index (source)

ParaEst-and-DistSel-by-ExtDist	Parameter-Estimation-and-Distribution-Selection-by-ExtDist (source)

Author

Haizhen Wu <h.wu2@massey.ac.nz>, A. Jonathan R. Godfrey <A.J.Godfrey@massey.ac.nz>, Kondaswamy Govindaraju <k.govindaraju@massey.ac.nz>, Sarah Pirikahu <s.pirikahu@massey.ac.nz>

Maintainer: Jonathan Godfrey <a.j.godfrey@massey.ac.nz>