The symmetric truncated normal distribution.

Density, distribution, quantile, random number generation and parameter estimation functions for the symmetric truncated normal distribution with parameters, sigma, a and b which represent the lower and upper truncation points respectively. Parameter estimation can be based on a weighted or unweighted i.i.d sample and can be carried out numerically.

dNormal_sym_trunc_ab(
  x,
  sigma = 0.3,
  a = 0,
  b = 1,
  params = list(sigma, a, b),
  ...
)

pNormal_sym_trunc_ab(
  q,
  sigma = 0.3,
  a = 0,
  b = 1,
  params = list(mu = 2, sigma = 5, a = 0, b = 1),
  ...
)

qNormal_sym_trunc_ab(
  p,
  sigma = 0.3,
  a = 0,
  b = 1,
  params = list(mu = 2, sigma = 5, a = 0, b = 1),
  ...
)

rNormal_sym_trunc_ab(
  n,
  mu = 2,
  sigma = 3,
  a = 0,
  b = 1,
  params = list(sigma, a, b),
  ...
)

eNormal_sym_trunc_ab(X, w, method = "numerical.MLE", ...)

lNormal_sym_trunc_ab(
  X,
  w,
  mu = 2,
  sigma = 3,
  a = 0,
  b = 1,
  params = list(sigma, a, b),
  logL = TRUE,
  ...
)

Arguments

x, q

A vector of quantiles.

a, b

Boundary parameters.

params

A list that includes all named parameters.

...

Additional parameters

p

A vector of probabilities.

n

Number of observations.

mu, sigma

Shape parameters.

X

Sample observations.

w

An optional vector of sample weights.

method

Parameter estimation method.

logL

logical;if TRUE, lNormal_sym_trunc_ab gives the log-likelihood, otherwise the likelihood is given.

Value

dNormal_sym_trunc_ab gives the density, pNormal_sym_trunc_ab the distribution function, qNormal_sym_trunc_ab the quantile function, rNormal_sym_trunc_ab generates random deviates,and eNormal_sym_trunc_ab estimates the parameters. lNormal_sym_trunc_ab provides the log-likelihood function.

Details

The normal symmetric truncated distribution is a special case of the trucated normal distribution. See Normal_trunc_ab.

See also

Author

Haizhen Wu and A. Jonathan R. Godfrey.