Distribution Selection Criteria.

A function to calculate the distribution selection criteria for a list of candidate fits.

DistSelCriteria(
  X,
  w = rep(1, length(X))/length(X),
  candDist = c("Beta_ab", "Laplace", "Normal"),
  criteria = c("logLik", "AIC", "AICc", "BIC", "MDL")
)

Arguments

X

Sample obersevations.

w

An optional vector of sample weights.

candDist

A vector of names of candidate distributions.

criteria

A vector of criteria to be calculated.

Value

An object of class matrix, containing the listed distribution selection criteria for the named distributions.

Details

When comparing models fitted by maximum likelihood to the same data, the smaller the AIC, BIC or MDL, the better the fit. When comparing models using the log-likelihood criterion, the larger the log-likelihood the better the fit.

Note

The MDL criterion only works for parameter estimation by numerical maximum likelihood.

Author

Haizhen Wu and A. Jonathan R. Godfrey.

Examples

Ozone <- airquality$Ozone
Ozone <- Ozone[!is.na(Ozone)] # Removing the NA's from Ozone data
DistSelCriteria(Ozone, candDist = c("Gamma", "Weibull", "Normal", "Exp"),
criteria = c("logLik","AIC","AICc", "BIC"))
#>        Gamma Weibull   Normal    Exp      
#> logLik -Inf  -542.6103 -569.6492 -549.9263
#> AIC    Inf   1089.221  1143.298  1101.853 
#> AICc   Inf   1089.327  1143.404  1101.888 
#> BIC    Inf   1094.728  1148.805  1104.606